Utilize rust-tools.nvim for using rust-analyzer.
Something about rust-analyzer wasn't working correctly (not updating
file cache/parsing on save, rename was not indexing correctly, and
more). rust-tools seems like it fixes the problem, and it's less on my
side to get working. If I feel like it in the future, I might pull the
settings into this repo and mess with them myself, but I double I'll do
that.